+++ src/main/java/eu/siacs/conversations/services/MessageSearchTask.java
@@ -47,6 +47,7 @@ import eu.siacs.conversations.entities.Message;
 import eu.siacs.conversations.entities.StubConversation;
 import eu.siacs.conversations.ui.interfaces.OnSearchResultsAvailable;
 import eu.siacs.conversations.utils.Cancellable;
+import eu.siacs.conversations.utils.MessageUtils;
 import eu.siacs.conversations.utils.ReplacingSerialSingleThreadExecutor;
 import rocks.xmpp.addr.Jid;
 
@@ -55,18 +56,18 @@ public class MessageSearchTask implements Runnable, Cancellable {
 	private static final ReplacingSerialSingleThreadExecutor EXECUTOR = new ReplacingSerialSingleThreadExecutor(MessageSearchTask.class.getName());
 
 	private final XmppConnectionService xmppConnectionService;
-	private final String term;
+	private final List<String> term;
 	private final OnSearchResultsAvailable onSearchResultsAvailable;
 
 	private boolean isCancelled = false;
 
-	private MessageSearchTask(XmppConnectionService xmppConnectionService, String term, OnSearchResultsAvailable onSearchResultsAvailable) {
+	private MessageSearchTask(XmppConnectionService xmppConnectionService, List<String> term, OnSearchResultsAvailable onSearchResultsAvailable) {
 		this.xmppConnectionService = xmppConnectionService;
 		this.term = term;
 		this.onSearchResultsAvailable = onSearchResultsAvailable;
 	}
 
-	public static void search(XmppConnectionService xmppConnectionService, String term, OnSearchResultsAvailable onSearchResultsAvailable) {
+	public static void search(XmppConnectionService xmppConnectionService, List<String> term, OnSearchResultsAvailable onSearchResultsAvailable) {
 		new MessageSearchTask(xmppConnectionService, term, onSearchResultsAvailable).executeInBackground();
 	}
 
@@ -87,25 +88,35 @@ public class MessageSearchTask implements Runnable, Cancellable {
 			final HashMap<String, Conversational> conversationCache = new HashMap<>();
 			final List<Message> result = new ArrayList<>();
 			cursor = xmppConnectionService.databaseBackend.getMessageSearchCursor(term);
+			long dbTimer = SystemClock.elapsedRealtime();
 			if (isCancelled) {
 				Log.d(Config.LOGTAG, "canceled search task");
 				return;
 			}
 			if (cursor != null && cursor.getCount() > 0) {
 				cursor.moveToLast();
+				final int indexBody = cursor.getColumnIndex(Message.BODY);
+				final int indexOob = cursor.getColumnIndex(Message.OOB);
+				final int indexConversation = cursor.getColumnIndex(Message.CONVERSATION);
+				final int indexAccount = cursor.getColumnIndex(Conversation.ACCOUNT);
+				final int indexContact = cursor.getColumnIndex(Conversation.CONTACTJID);
+				final int indexMode = cursor.getColumnIndex(Conversation.MODE);
 				do {
 					if (isCancelled) {
 						Log.d(Config.LOGTAG, "canceled search task");
 						return;
 					}
-					final String conversationUuid = cursor.getString(cursor.getColumnIndex(Message.CONVERSATION));
-					Conversational conversation;
-					if (conversationCache.containsKey(conversationUuid)) {
-						conversation = conversationCache.get(conversationUuid);
-					} else {
-						String accountUuid = cursor.getString(cursor.getColumnIndex(Conversation.ACCOUNT));
-						String contactJid = cursor.getString(cursor.getColumnIndex(Conversation.CONTACTJID));
-						int mode = cursor.getInt(cursor.getColumnIndex(Conversation.MODE));
+					final String body = cursor.getString(indexBody);
+					final boolean oob = cursor.getInt(indexOob) > 0;
+					if (MessageUtils.treatAsDownloadable(body,oob)) {
+						continue;
+					}
+					final String conversationUuid = cursor.getString(indexConversation);
+					Conversational conversation = conversationCache.get(conversationUuid);
+					if (conversation == null) {
+						String accountUuid = cursor.getString(indexAccount);
+						String contactJid = cursor.getString(indexContact);
+						int mode = cursor.getInt(indexMode);
 						conversation = findOrGenerateStub(conversationUuid, accountUuid, contactJid, mode);
 						conversationCache.put(conversationUuid, conversation);
 					}
@@ -114,7 +125,7 @@ public class MessageSearchTask implements Runnable, Cancellable {
 				} while (cursor.moveToPrevious());
 			}
 			long stopTimestamp = SystemClock.elapsedRealtime();
-			Log.d(Config.LOGTAG, "found " + result.size() + " messages in " + (stopTimestamp - startTimestamp) + "ms");
+			Log.d(Config.LOGTAG, "found " + result.size() + " messages in " + (stopTimestamp - startTimestamp) + "ms"+ " (db was "+(dbTimer - startTimestamp)+"ms)");
 			onSearchResultsAvailable.onSearchResultsAvailable(term, result);
 		} catch (Exception e) {
 			Log.d(Config.LOGTAG, "exception while searching ", e);