+++ src/main/java/eu/siacs/conversations/ui/ConversationFragment.java
@@ -85,6 +85,7 @@ import eu.siacs.conversations.ui.adapter.MessageAdapter;
 import eu.siacs.conversations.ui.util.ActivityResult;
 import eu.siacs.conversations.ui.util.AttachmentTool;
 import eu.siacs.conversations.ui.util.ConversationMenuConfigurator;
+import eu.siacs.conversations.ui.util.DateSeparator;
 import eu.siacs.conversations.ui.util.ListViewUtils;
 import eu.siacs.conversations.ui.util.MenuDoubleTabUtil;
 import eu.siacs.conversations.ui.util.PendingItem;
@@ -92,9 +93,12 @@ import eu.siacs.conversations.ui.util.PresenceSelector;
 import eu.siacs.conversations.ui.util.ScrollState;
 import eu.siacs.conversations.ui.util.SendButtonAction;
 import eu.siacs.conversations.ui.util.SendButtonTool;
+import eu.siacs.conversations.ui.util.ShareUtil;
 import eu.siacs.conversations.ui.widget.EditMessage;
+import eu.siacs.conversations.utils.GeoHelper;
 import eu.siacs.conversations.utils.MessageUtils;
 import eu.siacs.conversations.utils.NickValidityChecker;
+import eu.siacs.conversations.utils.Patterns;
 import eu.siacs.conversations.utils.QuickLoader;
 import eu.siacs.conversations.utils.StylingHelper;
 import eu.siacs.conversations.utils.TimeframeUtils;
@@ -790,14 +794,14 @@ public class ConversationFragment extends XmppFragment implements EditMessage.Ke
 				selectPresenceToAttachFile(choice);
 				break;
 			case ATTACHMENT_CHOICE_CHOOSE_IMAGE:
-				List<Uri> imageUris = AttachmentTool.extractUriFromIntent(data);
+				final List<Uri> imageUris = AttachmentTool.extractUriFromIntent(data);
 				for (Iterator<Uri> i = imageUris.iterator(); i.hasNext(); i.remove()) {
 					Log.d(Config.LOGTAG, "ConversationsActivity.onActivityResult() - attaching image to conversations. CHOOSE_IMAGE");
 					attachImageToConversation(conversation, i.next());
 				}
 				break;
 			case ATTACHMENT_CHOICE_TAKE_PHOTO:
-				Uri takePhotoUri = pendingTakePhotoUri.pop();
+				final Uri takePhotoUri = pendingTakePhotoUri.pop();
 				if (takePhotoUri != null) {
 					attachImageToConversation(conversation, takePhotoUri);
 				} else {
@@ -808,7 +812,7 @@ public class ConversationFragment extends XmppFragment implements EditMessage.Ke
 			case ATTACHMENT_CHOICE_RECORD_VIDEO:
 			case ATTACHMENT_CHOICE_RECORD_VOICE:
 				final List<Uri> fileUris = AttachmentTool.extractUriFromIntent(data);
-				String type = data.getType();
+				final String type = data == null ? null : data.getType();
 				final PresenceSelector.OnPresenceSelected callback = () -> {
 					for (Iterator<Uri> i = fileUris.iterator(); i.hasNext(); i.remove()) {
 						Log.d(Config.LOGTAG, "ConversationsActivity.onActivityResult() - attaching file to conversations. CHOOSE_FILE/RECORD_VOICE/RECORD_VIDEO");
@@ -1049,6 +1053,7 @@ public class ConversationFragment extends XmppFragment implements EditMessage.Ke
 			activity.getMenuInflater().inflate(R.menu.message_context, menu);
 			menu.setHeaderTitle(R.string.message_options);
 			MenuItem copyMessage = menu.findItem(R.id.copy_message);
+			MenuItem copyLink = menu.findItem(R.id.copy_link);
 			MenuItem quoteMessage = menu.findItem(R.id.quote_message);
 			MenuItem retryDecryption = menu.findItem(R.id.retry_decryption);
 			MenuItem correctMessage = menu.findItem(R.id.correct_message);
@@ -1062,6 +1067,13 @@ public class ConversationFragment extends XmppFragment implements EditMessage.Ke
 			if (!m.isFileOrImage() && !encrypted && !m.isGeoUri() && !m.treatAsDownloadable()) {
 				copyMessage.setVisible(true);
 				quoteMessage.setVisible(MessageUtils.prepareQuote(m).length() > 0);
+				String body = m.getMergedBody().toString();
+				if (ShareUtil.containsXmppUri(body)) {
+					copyLink.setTitle(R.string.copy_jabber_id);
+					copyLink.setVisible(true);
+				} else if (Patterns.AUTOLINK_WEB_URL.matcher(body).find()) {
+					copyLink.setVisible(true);
+				}
 			}
 			if (m.getEncryption() == Message.ENCRYPTION_DECRYPTION_FAILED) {
 				retryDecryption.setVisible(true);
@@ -1111,13 +1123,16 @@ public class ConversationFragment extends XmppFragment implements EditMessage.Ke
 	public boolean onContextItemSelected(MenuItem item) {
 		switch (item.getItemId()) {
 			case R.id.share_with:
-				shareWith(selectedMessage);
+				ShareUtil.share(activity, selectedMessage);
 				return true;
 			case R.id.correct_message:
 				correctMessage(selectedMessage);
 				return true;
 			case R.id.copy_message:
-				copyMessage(selectedMessage);
+				ShareUtil.copyToClipboard(activity, selectedMessage);
+				return true;
+			case R.id.copy_link:
+				ShareUtil.copyLinkToClipboard(activity, selectedMessage);
 				return true;
 			case R.id.quote_message:
 				quoteMessage(selectedMessage);
@@ -1126,7 +1141,7 @@ public class ConversationFragment extends XmppFragment implements EditMessage.Ke
 				resendMessage(selectedMessage);
 				return true;
 			case R.id.copy_url:
-				copyUrl(selectedMessage);
+				ShareUtil.copyUrlToClipboard(activity, selectedMessage);
 				return true;
 			case R.id.download_file:
 				startDownloadable(selectedMessage);
@@ -1496,7 +1511,7 @@ public class ConversationFragment extends XmppFragment implements EditMessage.Ke
 					intent = new Intent(getActivity(), RecordingActivity.class);
 					break;
 				case ATTACHMENT_CHOICE_LOCATION:
-					intent = new Intent(getActivity(), ShareLocationActivity.class);
+					intent = GeoHelper.getFetchIntent(activity);
 					break;
 			}
 			if (intent.resolveActivity(getActivity().getPackageManager()) != null) {
@@ -1570,43 +1585,6 @@ public class ConversationFragment extends XmppFragment implements EditMessage.Ke
 		builder.create().show();
 	}
 
-	private void shareWith(Message message) {
-		Intent shareIntent = new Intent();
-		shareIntent.setAction(Intent.ACTION_SEND);
-		if (message.isGeoUri()) {
-			shareIntent.putExtra(Intent.EXTRA_TEXT, message.getBody());
-			shareIntent.setType("text/plain");
-		} else if (!message.isFileOrImage()) {
-			shareIntent.putExtra(Intent.EXTRA_TEXT, message.getMergedBody().toString());
-			shareIntent.setType("text/plain");
-		} else {
-			final DownloadableFile file = activity.xmppConnectionService.getFileBackend().getFile(message);
-			try {
-				shareIntent.putExtra(Intent.EXTRA_STREAM, FileBackend.getUriForFile(getActivity(), file));
-			} catch (SecurityException e) {
-				Toast.makeText(getActivity(), activity.getString(R.string.no_permission_to_access_x, file.getAbsolutePath()), Toast.LENGTH_SHORT).show();
-				return;
-			}
-			shareIntent.addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION);
-			String mime = message.getMimeType();
-			if (mime == null) {
-				mime = "*/*";
-			}
-			shareIntent.setType(mime);
-		}
-		try {
-			startActivity(Intent.createChooser(shareIntent, getText(R.string.share_with)));
-		} catch (ActivityNotFoundException e) {
-			//This should happen only on faulty androids because normally chooser is always available
-			Toast.makeText(getActivity(), R.string.no_application_found_to_open_file, Toast.LENGTH_SHORT).show();
-		}
-	}
-
-	private void copyMessage(Message message) {
-		if (activity.copyTextToClipboard(message.getMergedBody().toString(), R.string.message)) {
-			Toast.makeText(getActivity(), R.string.message_copied_to_clipboard, Toast.LENGTH_SHORT).show();
-		}
-	}
 
 	private void deleteFile(Message message) {
 		if (activity.xmppConnectionService.getFileBackend().deleteFile(message)) {
@@ -1651,24 +1629,6 @@ public class ConversationFragment extends XmppFragment implements EditMessage.Ke
 			int size = messageList.size();
 			this.binding.messagesView.setSelection(size - 1);
 		});
-	}
-
-	private void copyUrl(Message message) {
-		final String url;
-		final int resId;
-		if (message.isGeoUri()) {
-			resId = R.string.location;
-			url = message.getBody();
-		} else if (message.hasFileOnRemoteHost()) {
-			resId = R.string.file_url;
-			url = message.getFileParams().url.toString();
-		} else {
-			url = message.getBody().trim();
-			resId = R.string.file_url;
-		}
-		if (activity.copyTextToClipboard(url, resId)) {
-			Toast.makeText(getActivity(), R.string.url_copied_to_clipboard, Toast.LENGTH_SHORT).show();
-		}
 	}
 
 	private void cancelTransmission(Message message) {
@@ -1943,6 +1903,7 @@ public class ConversationFragment extends XmppFragment implements EditMessage.Ke
 		final String downloadUuid = extras.getString(ConversationsActivity.EXTRA_DOWNLOAD_UUID);
 		final String text = extras.getString(ConversationsActivity.EXTRA_TEXT);
 		final String nick = extras.getString(ConversationsActivity.EXTRA_NICK);
+		final boolean asQuote = extras.getBoolean(ConversationsActivity.EXTRA_AS_QUOTE);
 		final boolean pm = extras.getBoolean(ConversationsActivity.EXTRA_IS_PRIVATE_MESSAGE, false);
 		if (nick != null) {
 			if (pm) {
@@ -1960,7 +1921,11 @@ public class ConversationFragment extends XmppFragment implements EditMessage.Ke
 				}
 			}
 		} else {
-			appendText(text);
+			if (text != null && asQuote) {
+				quoteText(text);
+			} else {
+				appendText(text);
+			}
 		}
 		final Message message = downloadUuid == null ? null : conversation.findMessageWithFileAndUuid(downloadUuid);
 		if (message != null) {
@@ -2159,13 +2124,7 @@ public class ConversationFragment extends XmppFragment implements EditMessage.Ke
 
 	protected void updateDateSeparators() {
 		synchronized (this.messageList) {
-			for (int i = 0; i < this.messageList.size(); ++i) {
-				final Message current = this.messageList.get(i);
-				if (i == 0 || !UIHelper.sameDay(this.messageList.get(i - 1).getTimeSent(), current.getTimeSent())) {
-					this.messageList.add(i, Message.createDateSeparator(current));
-					i++;
-				}
-			}
+			DateSeparator.addAll(this.messageList);
 		}
 	}