+++ src/main/java/eu/siacs/conversations/ui/SearchActivity.java
@@ -36,10 +36,14 @@ import android.text.Editable;
 import android.text.InputType;
 import android.text.TextWatcher;
 import android.util.Log;
+import android.view.ContextMenu;
 import android.view.Menu;
 import android.view.MenuItem;
+import android.view.View;
+import android.widget.AdapterView;
 import android.widget.EditText;
 
+import java.lang.ref.WeakReference;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -47,44 +51,96 @@ import eu.siacs.conversations.Config;
 import eu.siacs.conversations.R;
 import eu.siacs.conversations.databinding.ActivitySearchBinding;
 import eu.siacs.conversations.entities.Contact;
+import eu.siacs.conversations.entities.Conversation;
+import eu.siacs.conversations.entities.Conversational;
 import eu.siacs.conversations.entities.Message;
+import eu.siacs.conversations.entities.StubConversation;
 import eu.siacs.conversations.services.MessageSearchTask;
 import eu.siacs.conversations.ui.adapter.MessageAdapter;
 import eu.siacs.conversations.ui.interfaces.OnSearchResultsAvailable;
+import eu.siacs.conversations.ui.util.ChangeWatcher;
 import eu.siacs.conversations.ui.util.Color;
+import eu.siacs.conversations.ui.util.DateSeparator;
 import eu.siacs.conversations.ui.util.Drawable;
 import eu.siacs.conversations.ui.util.ListViewUtils;
+import eu.siacs.conversations.ui.util.PendingItem;
+import eu.siacs.conversations.ui.util.ShareUtil;
+import eu.siacs.conversations.utils.FtsUtils;
+import eu.siacs.conversations.utils.MessageUtils;
 
 import static eu.siacs.conversations.ui.util.SoftKeyboardUtils.hideSoftKeyboard;
 import static eu.siacs.conversations.ui.util.SoftKeyboardUtils.showKeyboard;
 
 public class SearchActivity extends XmppActivity implements TextWatcher, OnSearchResultsAvailable, MessageAdapter.OnContactPictureClicked {
 
+	private static final String EXTRA_SEARCH_TERM = "search-term";
+
 	private ActivitySearchBinding binding;
 	private MessageAdapter messageListAdapter;
 	private final List<Message> messages = new ArrayList<>();
+	private WeakReference<Message> selectedMessageReference = new WeakReference<>(null);
+	private final ChangeWatcher<List<String>> currentSearch = new ChangeWatcher<>();
+	private final PendingItem<String> pendingSearchTerm = new PendingItem<>();
+	private final PendingItem<List<String>> pendingSearch = new PendingItem<>();
 
 	@Override
-	public void onCreate(final Bundle savedInstanceState) {
-		super.onCreate(savedInstanceState);
+	public void onCreate(final Bundle bundle) {
+		final String searchTerm = bundle == null ? null : bundle.getString(EXTRA_SEARCH_TERM);
+		if (searchTerm != null) {
+			pendingSearchTerm.push(searchTerm);
+		}
+		super.onCreate(bundle);
 		this.binding = DataBindingUtil.setContentView(this, R.layout.activity_search);
 		setSupportActionBar((Toolbar) this.binding.toolbar);
 		configureActionBar(getSupportActionBar());
 		this.messageListAdapter = new MessageAdapter(this, this.messages);
 		this.messageListAdapter.setOnContactPictureClicked(this);
 		this.binding.searchResults.setAdapter(messageListAdapter);
+		registerForContextMenu(this.binding.searchResults);
 	}
 
 	@Override
 	public boolean onCreateOptionsMenu(final Menu menu) {
 		getMenuInflater().inflate(R.menu.activity_search, menu);
-		MenuItem searchActionMenuItem = menu.findItem(R.id.action_search);
-		EditText searchField = searchActionMenuItem.getActionView().findViewById(R.id.search_field);
+		final MenuItem searchActionMenuItem = menu.findItem(R.id.action_search);
+		final EditText searchField = searchActionMenuItem.getActionView().findViewById(R.id.search_field);
+		final String term = pendingSearchTerm.pop();
+		if (term != null) {
+			searchField.append(term);
+			List<String> searchTerm = FtsUtils.parse(term);
+			if (xmppConnectionService != null) {
+				if (currentSearch.watch(searchTerm)) {
+					xmppConnectionService.search(searchTerm, this);
+				}
+			} else {
+				pendingSearch.push(searchTerm);
+			}
+		}
 		searchField.addTextChangedListener(this);
 		searchField.setHint(R.string.search_messages);
-		searchField.setInputType(InputType.TYPE_CLASS_TEXT|InputType.TYPE_TEXT_FLAG_AUTO_COMPLETE);
-		showKeyboard(searchField);
+		searchField.setInputType(InputType.TYPE_CLASS_TEXT | InputType.TYPE_TEXT_FLAG_AUTO_COMPLETE);
+		if (term == null) {
+			showKeyboard(searchField);
+		}
 		return super.onCreateOptionsMenu(menu);
+	}
+
+	@Override
+	public void onCreateContextMenu(ContextMenu menu, View v, ContextMenu.ContextMenuInfo menuInfo) {
+		AdapterView.AdapterContextMenuInfo acmi = (AdapterView.AdapterContextMenuInfo) menuInfo;
+		final Message message = this.messages.get(acmi.position);
+		this.selectedMessageReference = new WeakReference<>(message);
+		getMenuInflater().inflate(R.menu.search_result_context, menu);
+		MenuItem copy = menu.findItem(R.id.copy_message);
+		MenuItem quote = menu.findItem(R.id.quote_message);
+		MenuItem copyUrl = menu.findItem(R.id.copy_url);
+		if (message.isGeoUri()) {
+			copy.setVisible(false);
+			quote.setVisible(false);
+		} else {
+			copyUrl.setVisible(false);
+		}
+		super.onCreateContextMenu(menu, v, menuInfo);
 	}
 
 	@Override
@@ -96,24 +152,77 @@ public class SearchActivity extends XmppActivity implements TextWatcher, OnSearc
 	}
 
 	@Override
+	public boolean onContextItemSelected(MenuItem item) {
+		final Message message = selectedMessageReference.get();
+		if (message != null) {
+			switch (item.getItemId()) {
+				case R.id.open_conversation:
+					switchToConversation(wrap(message.getConversation()));
+					break;
+				case R.id.share_with:
+					ShareUtil.share(this, message);
+					break;
+				case R.id.copy_message:
+					ShareUtil.copyToClipboard(this, message);
+					break;
+				case R.id.copy_url:
+					ShareUtil.copyUrlToClipboard(this, message);
+					break;
+				case R.id.quote_message:
+					quote(message);
+					break;
+			}
+		}
+		return super.onContextItemSelected(item);
+	}
+
+	@Override
+	public void onSaveInstanceState(Bundle bundle) {
+		List<String> term = currentSearch.get();
+		if (term != null && term.size() > 0) {
+			bundle.putString(EXTRA_SEARCH_TERM,FtsUtils.toUserEnteredString(term));
+		}
+		super.onSaveInstanceState(bundle);
+	}
+
+	private void quote(Message message) {
+		switchToConversationAndQuote(wrap(message.getConversation()), MessageUtils.prepareQuote(message));
+	}
+
+	private Conversation wrap(Conversational conversational) {
+		if (conversational instanceof Conversation) {
+			return (Conversation) conversational;
+		} else {
+			return xmppConnectionService.findOrCreateConversation(conversational.getAccount(),
+					conversational.getJid(),
+					conversational.getMode() == Conversational.MODE_MULTI,
+					true,
+					true);
+		}
+	}
+
+	@Override
 	protected void refreshUiReal() {
 
 	}
 
 	@Override
 	void onBackendConnected() {
-
+		final List<String> searchTerm = pendingSearch.pop();
+		if (searchTerm != null && currentSearch.watch(searchTerm)) {
+			xmppConnectionService.search(searchTerm, this);
+		}
 	}
 
 	private void changeBackground(boolean hasSearch, boolean hasResults) {
 		if (hasSearch) {
 			if (hasResults) {
-				binding.searchResults.setBackgroundColor(Color.get(this,R.attr.color_background_secondary));
+				binding.searchResults.setBackgroundColor(Color.get(this, R.attr.color_background_secondary));
 			} else {
-				binding.searchResults.setBackground(Drawable.get(this,R.attr.activity_background_no_results));
+				binding.searchResults.setBackground(Drawable.get(this, R.attr.activity_background_no_results));
 			}
 		} else {
-			binding.searchResults.setBackground(Drawable.get(this,R.attr.activity_background_search));
+			binding.searchResults.setBackground(Drawable.get(this, R.attr.activity_background_search));
 		}
 	}
 
@@ -129,21 +238,27 @@ public class SearchActivity extends XmppActivity implements TextWatcher, OnSearc
 
 	@Override
 	public void afterTextChanged(Editable s) {
-		String term = s.toString().trim();
-		if (term.length() > 0) {
-			xmppConnectionService.search(s.toString().trim(), this);
+		final List<String> term = FtsUtils.parse(s.toString().trim());
+		if (!currentSearch.watch(term)) {
+			return;
+		}
+		if (term.size() > 0) {
+			xmppConnectionService.search(term, this);
 		} else {
 			MessageSearchTask.cancelRunningTasks();
 			this.messages.clear();
+			messageListAdapter.setHighlightedTerm(null);
 			messageListAdapter.notifyDataSetChanged();
 			changeBackground(false, false);
 		}
 	}
 
 	@Override
-	public void onSearchResultsAvailable(String term, List<Message> messages) {
+	public void onSearchResultsAvailable(List<String> term, List<Message> messages) {
 		runOnUiThread(() -> {
 			this.messages.clear();
+			messageListAdapter.setHighlightedTerm(term);
+			DateSeparator.addAll(messages);
 			this.messages.addAll(messages);
 			messageListAdapter.notifyDataSetChanged();
 			changeBackground(true, messages.size() > 0);