+++ src/main/java/eu/siacs/conversations/ui/interfaces/OnSearchResultsAvailable.java
@@ -35,6 +35,6 @@ import eu.siacs.conversations.entities.Message;
 
 public interface OnSearchResultsAvailable {
 
-	void onSearchResultsAvailable(String term, List<Message> messages);
+	void onSearchResultsAvailable(List<String> term, List<Message> messages);
 
 }