+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+### Version 2.2.0
+* Integrate Voice Recorder
+* Integrate Share Location
+* Added ability to search messages
+
+### Version 2.1.4
+* bug fixes
+
+### Version 2.1.3
+* Do not process stanzas with invalid JIDs
+
 ### Version 2.1.2
 * Fixed avatars not being displayed on new installs